backend/pyproject.toml
[project]
name = "emotion-ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "openai-whisper>=20240930",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.2.1",
    "sentencepiece>=0.2.1",
    "supabase>=2.28.0",
    "torch>=2.10.0",
    "transformers>=5.2.0",
    "tzdata>=2024.1",
    "uvicorn>=0.41.0",
]
